UpdatesRemove unused views and restructure mode components
Removed 'PartySearcherView' and 'TestView' as they were not being utilised in the project. Also updated the routing paths in 'index.ts' to reflect the changes. Moreover, the 'CardMode', 'PureMode', and 'DefaultMode' components from the root directory were moved to the 'src/views' directory for better project organisation and consistency. Paths in 'App.vue' and 'DefaultMode.vue' were updated accordingly to ensure proper functioning of the application.