Update file structure and improve UI design
This commit includes a significant restructuring of the 'calculator' feature files in the application. The CalculatorView component has been moved inside its own folder within the 'views' directory to better organize the project structure. Additionally, within the Router index file, the path to import the CalculatorView component was updated accordingly.

UI changes were made within the CharacterWikiCard and CalculatorPanel components. The tag within the CharacterWikiCard component was changed from 'v-card' to 'div' to streamline the design. Also, the 'elevation' attribute within the CalculatorPanel component was reduced for a more subtle shadow effect, and padding was removed to maximize space for content display.