Refactor character handling in Worldflipper module
Changed the way characters are handled for consistency and cleaner code. Before, characters were specific objects, but now we use a more general 'obj' which can be of various types, including characters and equipment, annotated as a WorldflipperObject. This change allows us to use the same variable for all objects of the worldflipper module, reducing the complexity in handling different cases. Also, it improves the readability of the components using this 'obj', as they no longer need to handle separate cases for characters, equipment, and possibly other types in the future. Therefore, it is a step forward in maintainability and scalability of the code.