Remove unused Icon components and add new EmptyIcon component
This commit removes several unused Vue icon components (IconCommunity, IconDocumentation, IconEcosystem, IconSupport, IconTooling) and adds a new EmptyIcon component. These changes are made to improve code maintainability and to add new functionality needed for the calculator panel in the Vue application. The new EmptyIcon component is used to signify a slot where an icon can be placed but currently is not, in the calculator panel. The removed Icon components were not in use which lead to clutter in the codebase. This modification ensures our codebase is clean and only contains components that are used.